HH-65: Исправление замечаний detect
