--- app/src/main/java/ru/practicum/android/diploma/features/filters/presentation/view_model/SearchFilterViewModel.kt
@@ -1,50 +1,0 @@
-package ru.practicum.android.diploma.features.filters.presentation.view_model
-
-import androidx.lifecycle.LiveData
-import androidx.lifecycle.MutableLiveData
-import androidx.lifecycle.ViewModel
-import ru.practicum.android.diploma.features.filters.domain.model.Filter
-
-class SearchFilterViewModel() : ViewModel() {
-
-    private val _currentFilter = MutableLiveData<Filter>(Filter())
-    val currentFilter: LiveData<Filter> = _currentFilter
-
-    val latestSearchFilter = Filter()
-
-    fun getIndustries() {
-        //Метод возвращающий "Отрасль"
-    }
-
-    fun setIndustry(industry: IndustryUI?) {
-        //Метод заменяющий отрасль
-    }
-
-    fun setChosenCountry(country: CountryUI?) {
-        //Метод заменяющий страну
-    }
-
-    fun setChosenRegion(region: RegionUI?) {
-        //Метод заменяющий регион
-    }
-
-    fun setSalary(salary: Int?) {
-        //Метод заменяющий сумму ожидаемой зарплаты
-    }
-
-    fun clearPlaceOfWork() {
-        //Метод очищающий граф "Место работы"
-    }
-
-    fun saveFilter() {
-        //Метод сохраняющий состояние "фильтра"
-    }
-
-    fun retrySearchQueryWithFilterSearch() {
-        //Метод повторяющий поисковый запрос
-    }
-
-    fun clearFilter() {
-        //Метод сбрасывающий настройки фильтра
-    }
-}
